+++ AdventOfCode2020/day15/day15.cpp
@@ -31,20 +31,20 @@ void Process(std::vector<int> input)
 	printf("  - Given the starting numbers %s the 2020th number spoken is %d.\n", stringInput.data(), output.back());
 }
 
-void Process2(std::vector<__int64> input, __int64 target = 2020L)
+void Process2(std::vector<int64> input, int64 target = 2020L)
 {
-	std::map<__int64, __int64> seen;
-	__int64 ix = 0;
-	for (__int64 i : input)
+	std::map<int64, int64> seen;
+	int64 ix = 0;
+	for (int64 i : input)
 	{
-		if (ix < input.size() - 1)
+		if (ix < (int64)input.size() - 1)
 			seen[i] = ix;
 		++ix;
 	}
-	__int64 number = input.back();
+	int64 number = input.back();
 	while (ix < target)
 	{
-		__int64 newnum;
+		int64 newnum;
 		if (!seen.count(number))
 		{
 			newnum = 0;
@@ -56,11 +56,9 @@ void Process2(std::vector<__int64> input, __int64 target = 2020L)
 		seen[number] = ix-1;
 		++ix;
 		number = newnum;
-//		if (count > max)
-//			max = count;
 	}
 	std::string stringInput;
-	for (__int64 i : input)
+	for (int64 i : input)
 	{
 		stringInput += std::to_string(i);
 		stringInput += ',';
@@ -79,7 +77,7 @@ int main()
 	Process2({ 3,1,2 });
 	Process2({ 13,16,0,12,15,1 });
 	printf("\n");
-	__int64 target = 30000000L;
+	int64 target = 30000000L;
 	Process2({ 0,3,6 }, target);
 	Process2({ 1,3,2 }, target);
 	Process2({ 2,1,3 }, target);
